Update to the test fetch
I have figured out that response.json() formats the information we want into a json, and then when we save that file using the setData(data) hook, data will also be set to a JSON file, we can read from this JSON file conditionally by structuring how we parse the JSON file in the output. For what I did, I took data[0], and since I am using the guardians table that gives me the first guardian on the list. By saying data[0].guardianFirstName it returns that corresponding value.